+++ src/main/java/fuzs/goldenagecombat/client/handler/AttributesTooltipHandler.java
@@ -3,8 +3,11 @@ package fuzs.goldenagecombat.client.handler;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Multimap;
 import fuzs.goldenagecombat.GoldenAgeCombat;
+import fuzs.goldenagecombat.handler.AttackAttributeHandler;
+import fuzs.goldenagecombat.registry.ModRegistry;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.entity.EquipmentSlot;
@@ -13,11 +16,13 @@ import net.minecraft.world.entity.ai.attributes.Attribute;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import org.apache.commons.compress.utils.Lists;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.Map;
@@ -35,13 +40,102 @@ public class AttributesTooltipHandler {
             if (!GoldenAgeCombat.CONFIG.client().tooltip.removeAllAttributes && startIndex != -1 && GoldenAgeCombat.CONFIG.client().tooltip.oldAttributes) {
                 this.addOldStyleAttributes(list, evt.getItemStack(), evt.getPlayer(), startIndex);
             }
-        } else if (GoldenAgeCombat.CONFIG.server().classic.removeCooldown) {
-            this.removeAttackSpeedAttribute(list);
+            return;
+        }
+        if (GoldenAgeCombat.CONFIG.server().classic.removeCooldown) {
+            this.removeAttribute(list, Attributes.ATTACK_SPEED);
+        }
+        if (GoldenAgeCombat.CONFIG.server().attributes.increasedAttackReach) {
+            this.replaceOrAddDefaultAttribute(list, ModRegistry.ATTACK_REACH_ATTRIBUTE.get(), AttackAttributeHandler.BASE_ATTACK_REACH_UUID, evt.getItemStack(), evt.getPlayer());
+        } else {
+            this.removeAttribute(list, ModRegistry.ATTACK_REACH_ATTRIBUTE.get());
+        }
+        if (evt.getItemStack().getItem() instanceof ArmorItem) {
+
         }
     }
 
-    private void removeAttackSpeedAttribute(List<Component> list) {
-        list.removeIf(component -> component.toString().contains("attribute.name.generic.attack_speed"));
+    private void removeAttribute(List<Component> list, Attribute attribute) {
+        list.removeIf(component -> this.compareToAttributeComponent(attribute, null, component));
+    }
+
+    private boolean compareToAttributeComponent(Attribute attribute, @Nullable AttributeModifier attributemodifier, Component component) {
+        TranslatableComponent translatableComponent = null;
+        if (component instanceof TranslatableComponent) {
+            translatableComponent = (TranslatableComponent) component;
+        } else if (component instanceof MutableComponent mutableComponent && !mutableComponent.getSiblings().isEmpty() && mutableComponent.getSiblings().get(0) instanceof TranslatableComponent) {
+            translatableComponent = (TranslatableComponent) mutableComponent.getSiblings().get(0);
+        }
+        if (translatableComponent != null) {
+            double scaledAmount = 0.0;
+            String translationKey = null;
+            if (attributemodifier != null) {
+                double attributeAmount = attributemodifier.getAmount();
+                scaledAmount = this.getScaledAttributeAmount(attributeAmount, attribute, attributemodifier);
+                if (attributeAmount > 0.0D) {
+                    translationKey = "attribute.modifier.plus." + attributemodifier.getOperation().toValue();
+                } else if (attributeAmount < 0.0D) {
+                    scaledAmount *= -1.0D;
+                    translationKey = "attribute.modifier.take." + attributemodifier.getOperation().toValue();
+                }
+            }
+            if ((attributemodifier == null || translationKey != null && translatableComponent.getKey().equals(translationKey)) && translatableComponent.getArgs().length >= 2) {
+                final Object[] args = translatableComponent.getArgs();
+                if ((attributemodifier == null || args[0].equals(ItemStack.ATTRIBUTE_MODIFIER_FORMAT.format(scaledAmount))) && args[1] instanceof TranslatableComponent translatableComponent1) {
+                    return translatableComponent1.getKey().equals(attribute.getDescriptionId());
+                }
+            }
+        }
+        return false;
+    }
+
+    private void replaceOrAddDefaultAttribute(List<Component> list, Attribute attribute, UUID attributeId, ItemStack stack, @Nullable Player player) {
+        final Map<EquipmentSlot, Multimap<Attribute, AttributeModifier>> map = this.getSlotToAttributeMap(stack);
+        for (Map.Entry<EquipmentSlot, Multimap<Attribute, AttributeModifier>> slotToAttributeMap : map.entrySet()) {
+            AttributeModifier attributeModifier = null;
+            for (Map.Entry<Attribute, AttributeModifier> attributeToModifier : slotToAttributeMap.getValue().entries()) {
+                if (attributeToModifier.getKey().equals(attribute) && attributeToModifier.getValue().getId().equals(attributeId)) {
+                    attributeModifier = attributeToModifier.getValue();
+                    break;
+                }
+            }
+            if (attributeModifier != null) {
+                final double attributeBaseAmount = attributeModifier.getAmount();
+                double attributeAmount = attributeBaseAmount;
+                if (player != null) {
+                    attributeAmount += player.getAttributeBaseValue(attribute);
+                }
+                double scaledAmount = this.getScaledAttributeAmount(attributeAmount, attribute, attributeModifier);
+                for (int i = 0; i < list.size(); i++) {
+                    final Component component = list.get(i);
+                    if (attributeBaseAmount != 0.0) {
+                        if (this.compareToAttributeComponent(attribute, attributeModifier, component)) {
+                            list.set(i, new TextComponent(" ").append(new TranslatableComponent("attribute.modifier.equals." + attributeModifier.getOperation().toValue(), ItemStack.ATTRIBUTE_MODIFIER_FORMAT.format(scaledAmount), new TranslatableComponent(attribute.getDescriptionId()))).withStyle(ChatFormatting.DARK_GREEN));
+                            break;
+                        }
+                    } else {
+                        if (component instanceof TranslatableComponent translatableComponent && translatableComponent.getKey().equals("item.modifiers." + slotToAttributeMap.getKey().getName())) {
+                            list.add(++i, new TextComponent(" ").append(new TranslatableComponent("attribute.modifier.equals." + attributeModifier.getOperation().toValue(), ItemStack.ATTRIBUTE_MODIFIER_FORMAT.format(scaledAmount), new TranslatableComponent(attribute.getDescriptionId()))).withStyle(ChatFormatting.DARK_GREEN));
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private double getScaledAttributeAmount(double attributeAmount, Attribute attribute, AttributeModifier attributemodifier) {
+        double d1;
+        if (attributemodifier.getOperation() != AttributeModifier.Operation.MULTIPLY_BASE && attributemodifier.getOperation() != AttributeModifier.Operation.MULTIPLY_TOTAL) {
+            if (attribute.equals(Attributes.KNOCKBACK_RESISTANCE)) {
+                d1 = attributeAmount * 10.0D;
+            } else {
+                d1 = attributeAmount;
+            }
+        } else {
+            d1 = attributeAmount * 100.0D;
+        }
+        return d1;
     }
 
     private int removeAllAttributes(List<Component> list) {
@@ -61,7 +155,7 @@ public class AttributesTooltipHandler {
         return -1;
     }
 
-    private void addOldStyleAttributes(List<Component> list, ItemStack stack, Player player, int startIndex) {
+    private void addOldStyleAttributes(List<Component> list, ItemStack stack, @Nullable Player player, int startIndex) {
         final Map<EquipmentSlot, Multimap<Attribute, AttributeModifier>> map = this.getSlotToAttributeMap(stack);
         if (map.size() == 1) {
             List<Component> tmpList = Lists.newArrayList();
@@ -132,9 +226,10 @@ public class AttributesTooltipHandler {
         return map;
     }
 
-    private void addAttributesToTooltip(List<Component> list, Player player, ItemStack stack, Multimap<Attribute, AttributeModifier> multimap) {
+    private void addAttributesToTooltip(List<Component> list, @Nullable Player player, ItemStack stack, Multimap<Attribute, AttributeModifier> multimap) {
         for (Map.Entry<Attribute, AttributeModifier> entry : multimap.entries()) {
             if (GoldenAgeCombat.CONFIG.server().classic.removeCooldown && entry.getKey().equals(Attributes.ATTACK_SPEED)) continue;
+            if (!GoldenAgeCombat.CONFIG.server().attributes.increasedAttackReach && entry.getKey().equals(ModRegistry.ATTACK_REACH_ATTRIBUTE.get())) continue;
             AttributeModifier attributemodifier = entry.getValue();
             double d0 = attributemodifier.getAmount();
             boolean flag = false;
@@ -145,6 +240,9 @@ public class AttributesTooltipHandler {
                     flag = true;
                 } else if (attributemodifier.getId().equals(BASE_ATTACK_SPEED_UUID)) {
                     d0 += player.getAttributeBaseValue(Attributes.ATTACK_SPEED);
+                    flag = true;
+                } else if (attributemodifier.getId().equals(AttackAttributeHandler.BASE_ATTACK_REACH_UUID)) {
+                    d0 += player.getAttributeBaseValue(ModRegistry.ATTACK_REACH_ATTRIBUTE.get());
                     flag = true;
                 }
             }